Refactor tests, use some abstract-leveldown tests
