+++ package.json
@@ -30,7 +30,6 @@
     "levelup": "^1.3.1",
     "rc": "^1.1.6",
     "standard": "^6.0.7",
-    "tape": "^4.4.0",
-    "xtend": "^4.0.1"
+    "tape": "^4.4.0"
   }
 }
