+++ test/abstract-leveldown.js
@@ -1,21 +1,25 @@
+const test = require('tape')
+const util = require('./util')
+const pgdown = require('../')
+
 // compatibility w/ leveldown api
 
-// require('abstract-leveldown/abstract/leveldown-test').args(factory, test, testCommon)
+require('abstract-leveldown/abstract/leveldown-test').args(pgdown, test, util)
 
-// require('abstract-leveldown/abstract/open-test').args(factory, test, testCommon)
-// require('abstract-leveldown/abstract/open-test').open(factory, test, testCommon)
+require('abstract-leveldown/abstract/open-test').args(pgdown, test, util)
+require('abstract-leveldown/abstract/open-test').open(pgdown, test, util)
 
-// require('abstract-leveldown/abstract/put-test').all(factory, test, testCommon)
+require('abstract-leveldown/abstract/put-test').all(pgdown, test, util)
 
-// require('abstract-leveldown/abstract/del-test').all(factory, test, testCommon)
+require('abstract-leveldown/abstract/del-test').all(pgdown, test, util)
 
-// require('abstract-leveldown/abstract/get-test').all(factory, test, testCommon)
+require('abstract-leveldown/abstract/get-test').all(pgdown, test, util)
 
-// require('abstract-leveldown/abstract/put-get-del-test').all(factory, test, testCommon, testBuffer)
+// require('abstract-leveldown/abstract/put-get-del-test').all(pgdown, test, util, testBuffer)
 
-// require('abstract-leveldown/abstract/iterator-test').all(factory, test, testCommon)
+// require('abstract-leveldown/abstract/iterator-test').all(pgdown, test, util)
 
-// require('abstract-leveldown/abstract/batch-test').all(factory, test, testCommon)
-// require('abstract-leveldown/abstract/chained-batch-test').all(factory, test, testCommon)
+// require('abstract-leveldown/abstract/batch-test').all(pgdown, test, util)
+// require('abstract-leveldown/abstract/chained-batch-test').all(pgdown, test, util)
 
-// require('abstract-leveldown/abstract/close-test').close(factory, test, testCommon)
+require('abstract-leveldown/abstract/close-test').close(pgdown, test, util)
